requirements.txt
django
django-crispy_forms
django-braces
djangorestframework
markdown
django-filter
django-activity-stream